+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.2.0 - 5 March 2020
+
+- Fix: Handle 'successful' API errors (i.e., HTTP 200 response, but API error such as User Not Found)
+- Documented expansion fields
+- Documented methods
+
 # 0.1.0 - 5 March 2020
 
 Initial gem commit